Rename function and add docstring
The function `_is_equal_source` is renamed to
`_contains_commented_out_code`. This change reflects the larger scope
that the function now has.

The function was previously only checking the difference between the
filtered source and the physical line.

Now the function also needs to test if the physical line even contains
a comment before the line is reduced by the eradicate function.
This scope is a bit larger which is reflected in the new name.

The previous name `_is_equal_source` was derived from the inner
implementation. To me it was not immediately clear why we are testing
for "source equality". The new name `_contains_commented_out_code`
describes what the function can tell us, but now how it is implemented.
The implementation is outside of the concern of the caller. All that
the caller is interested in is, if the current physical line contains
a comment or not.

**Attention** With the change of the name, the logic is also inverted.
While `_is_equal_source` was `True` when the original and the filtered
line are the same, `_contains_commented_out_code` is `True` when they
are different (because the commented code was removed by eradicate).